% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsearch.R
\name{libris_searchbox_html}
\alias{libris_searchbox_html}
\title{HTML for a LIBRIS search box that can be integrated in HTML (rmarkdown docs, etc)}
\usage{
libris_searchbox_html()
}
\value{
character string with HTML snippet
}
\description{
HTML for a LIBRIS search box that can be integrated in HTML (rmarkdown docs, etc)
}
