% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsearch.R
\name{xsearch}
\alias{xsearch}
\title{Retrieve records from LIBRIS Xsearch API}
\usage{
xsearch(
  query = NULL,
  start = 1,
  n = 10,
  order = c("rank", "alphabetical", "-alphabetical", "chronological", "-chronological"),
  format_level = c("brief", "full"),
  holdings = FALSE,
  database = c("libris", "swepub"),
  api_config = NULL
)
}
\arguments{
\item{query}{text query as defined by the custom query language described here \url{http://librishelp.libris.kb.se/help/search_language_swe.jsp}

Operators are: \verb{AND, OR, NOT, ADJ, SAME, XOR}

A short list of examples follow below.
\itemize{
\item author is Ludwig Wittgenstein: \code{forf:(Ludwig+Wittgenstein)}
\item publications with title "Röda rummet": \code{TIT:"Röda rummet"}
\item publications with a set of titles (experiment or experiments or experimental): \verb{TIT=experiment OR TIT:experiments OR TIT:experimental}
\item publications by Horatius Quintus: \verb{PERS:(Horatius Quintus)}
}}

\item{start}{integer that specifies the starting position in a hitlist from which records should be taken}

\item{n}{the number of records to be retrieved, to a maximum of 200}

\item{order}{sort order, one of "rank", "alphabetical", "-alphabetical",
"chronological", "-chronological" where an opening hyphen stands for sorting in reverse order}

\item{format_level}{one of "brief" or "full", which specifies whether “see” or ”see also” links to authority forms (9XX fields) should be included.}

\item{holdings}{boolean defaulting to FALSE, specifies whether holding information should be included. Works only with the marcxml and mods formats.}

\item{database}{one of "libris" or "swepub"}

\item{api_config}{a configuration setting for the search API including base URL etc, by default from config()}
}
\value{
results records returned from the search
}
\description{
The database used can be specified - the default is to use LIBRIS but "swepub" is also available
}
\examples{
\dontrun{
xsearch(query = "WAMK:\"film noir\"")
xsearch(query = "PERS:(vitruvius) SPR:swe")

xsearch(query = "stolpverk SPR:swe", database = "swepub")$content \%>\%
dplyr::slice(1) \%>\%
dplyr::glimpse()
}
}
