% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsearch.R
\name{xsearch_crawl}
\alias{xsearch_crawl}
\title{Retrieve more records from LIBRIS Xsearch API in batches}
\usage{
xsearch_crawl(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=xsearch]{xsearch}}
  \describe{
    \item{\code{query}}{text query as defined by the custom query language described here \url{http://librishelp.libris.kb.se/help/search_language_swe.jsp}

Operators are: \verb{AND, OR, NOT, ADJ, SAME, XOR}

A short list of examples follow below.
\itemize{
\item author is Ludwig Wittgenstein: \code{forf:(Ludwig+Wittgenstein)}
\item publications with title "Röda rummet": \code{TIT:"Röda rummet"}
\item publications with a set of titles (experiment or experiments or experimental): \verb{TIT=experiment OR TIT:experiments OR TIT:experimental}
\item publications by Horatius Quintus: \verb{PERS:(Horatius Quintus)}
}}
    \item{\code{start}}{integer that specifies the starting position in a hitlist from which records should be taken}
    \item{\code{n}}{the number of records to be retrieved, to a maximum of 200}
    \item{\code{order}}{sort order, one of "rank", "alphabetical", "-alphabetical",
"chronological", "-chronological" where an opening hyphen stands for sorting in reverse order}
    \item{\code{format_level}}{one of "brief" or "full", which specifies whether “see” or ”see also” links to authority forms (9XX fields) should be included.}
    \item{\code{holdings}}{boolean defaulting to FALSE, specifies whether holding information should be included. Works only with the marcxml and mods formats.}
    \item{\code{database}}{one of "libris" or "swepub"}
    \item{\code{api_config}}{a configuration setting for the search API including base URL etc, by default from config()}
  }}
}
\value{
results records returned from the search
}
\description{
This function retrieves batches of record (paging through results) for
between 200 and 10000 records in one go, displays a progress bar while progressing
and then returns all results.
}
\details{
The database used can be specified - the default is to use LIBRIS but "swepub" is also available
}
\examples{
\dontrun{
xsearch_crawl(query = "corona", database = "swepub")
}
}
