% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-xsearch.R
\docType{package}
\name{XsearchLIBRIS}
\alias{XsearchLIBRIS}
\alias{package-XsearchLIBRIS}
\title{XsearchLIBRIS}
\description{
LIBRIS is an online search service of the National Library of Sweden.
Search national, academic, and several public libraries of Sweden for
books, periodicals, articles, maps, posters, printed music, electronic
resources, and more.
}
\keyword{package}
